favorites page working
