+++ schema.sql
@@ -1,3 +1,11 @@
+-- Heroku setup
 DROP DATABASE IF EXISTS wh9vaw41o85ds5ll;
 
 CREATE DATABASE wh9vaw41o85ds5ll;
+
+-- Local database setup
+DROP DATABASE IF EXISTS restaurant_db;
+
+CREATE DATABASE restaurant_db;
+
+USE restaurant_db;
