+++ views/favorites.handlebars
@@ -37,9 +37,6 @@
           <div class="content">
               <h1 class="header">\{{this.name}}</h1>
           </div>
-          <div class="extra content">
-            <div class="rating">Rating: \{{this.rating}}/5</div>
-          </div>
           <div class="content info">
             <a class="yelp-link" href=\{{this.yelp_url}}>Get Info</a> 
           </div>
@@ -50,9 +47,9 @@
 <script src="https://code.jquery.com/jquery.js"></script>
 <script>
     // Displays the card handlebar
-displayFavoritesAPI = function() {
-    return $.get("/liked").then(function(res){
-        $.get("/liked/restaurants").then(function(result){
+displayFavorites = function() {
+    return $.get("/favorites").then(function(res){
+        $.get("/favorites/restaurants").then(function(result){
             for (i=0;i<result.length;i++){
                 console.log(result[i]);
             }
@@ -70,5 +67,5 @@ displayFavoritesAPI = function() {
         })
     });
 };
-displayFavoritesAPI();
+displayFavorites();
 </script>
